Add aarch64 support for PiDP11
I have a stack of Raspberry Pi 3s; I'm planning to use one for a PiDP-11
replica. The Pi 3 is 64-bit capable, and that's the Rasbpian default as
of earlier this year.

Add the option to build the PiDP server with a 64-bit (aarch64)
toolchain.

I haven't tested this yet, so caveat emptor!